requirements.txt
aiosfstream >= 0.5.0
dnspython >= 2.0.0
jsonpickle >= 1.4.1
pymongo >= 3.11.0
requests >= 2.24.0
requests-toolbelt >= 0.9.1
simple-salesforce >= 1.10.1
PyJWT >= 1.7.1
beautifulsoup4 >= 4.9.3
lxml >= 4.6.1
jira >= 2.0.0
aiohttp >= 3.7.2
zenpy >= 2.0.22
kubernetes >= 12.0.0
pycryptodome >= 3.9.9
cryptography~=3.3.1
